Rename to constants.rs
