+++ crates/icon/src/lib.rs
@@ -11,12 +11,12 @@ pub const FOLDER_ICON: &str = "";
 pub const DEFAULT_FILER_ICON: &str = "";
 
 // The code is generated by ../update_icon_constants.py
+//
+// constants.rs declares these two variables:
+//
 // pub static ref EXACTMATCH_MAP: HashMap<&'static str, &'static str>
-include!("exactmatch_map.rs");
-
-// The code is generated by ../update_icon_constants.py
 // pub static ref EXTENSION_MAP: HashMap<&'static str, &'static str>
-include!("extension_map.rs");
+include!("constants.rs");
 
 #[inline]
 pub fn icon_for_path(path: &Path) -> &str {
